changes controllers & validationfiles

+++ Controllers/SongController/songController.js
@@ -1,27 +1,38 @@
 const Models = require("../../models/index");
-const { http, messages } = require("../../constant/constant");
-
-const {
-  validateAddSongs,
-} = require("../../services/validations/songValidation");
+const { http, messages, role } = require("../../constant/constant");
+const { validateSongs } = require("../../services/validations/songValidation");
 
 module.exports.addSong = async (req, res) => {
   try {
-    const { value } = validateAddSongs(req.body, res);
-    const song = await Models.Song.create({
-      song_name: value.song_name,
-      singer: value.singer,
-      creator_id: value.creator_id,
+    const userId = req.userId;
+    const { value } = validateSongs(req.body, res);
+    const isAdmin = await Models.User.findOne({
+      where: { id: userId },
     });
 
-    await song.addGenres(value.genres);
-    res.status(http.OK.code).send({
-      success: true,
-      data: song.dataValues.song_name,
-      message: http.OK.message,
-    });
+    if (isAdmin.dataValues.user_type === role.admin) {
+      const song = await Models.Song.create({
+        song_name: value.song_name,
+        singer: value.singer,
+        creator_id: userId,
+      });
+
+      await song.addGenres(value.genres);
+      res.status(http.OK.code).send({
+        success: true,
+        data: song.dataValues.song_name,
+        message: http.OK.message,
+      });
+    } else {
+      res.status(http.FORBIDDEN.code).send({
+        success: false,
+        data: null,
+        message: http.FORBIDDEN.message,
+      });
+    }
   } catch (e) {
     res.status(http.INTERNAL_SERVER_ERROR.code).send({
+      success: false,
       data: null,
       message: http.INTERNAL_SERVER_ERROR.message,
     });
@@ -30,15 +41,30 @@ module.exports.addSong = async (req, res) => {
 
 module.exports.deleteSong = async (req, res) => {
   try {
+    const userId = req.userId;
     const id = parseInt(req.params.id);
-    await Models.Song.destroy({
-      where: {
-        id,
-      },
+
+    const isAdmin = await Models.User.findOne({
+      where: { id: userId },
     });
-    res.send({
-      message: messages.REMOVE,
-    });
+
+    if (isAdmin.dataValues.user_type === role.admin) {
+      await Models.Song.destroy({
+        where: {
+          id,
+        },
+      });
+      res.status(http.OK.code).send({
+        success: true,
+        message: messages.REMOVED,
+      });
+    } else {
+      res.status(http.FORBIDDEN.code).send({
+        success: false,
+        data: null,
+        message: http.FORBIDDEN.message,
+      });
+    }
   } catch (e) {
     res.status(http.INTERNAL_SERVER_ERROR.code).send({
       success: false,
@@ -50,33 +76,50 @@ module.exports.deleteSong = async (req, res) => {
 
 module.exports.UpdateSongData = async (req, res) => {
   try {
+    const userId = req.userId;
     const id = parseInt(req.params.id);
-    const { name, meta, singer } = req.body;
-
+    const { value } = validateSongs(req.body);
     const updateFields = {
-      song_name: name,
-      meta,
-      singer,
+      song_name: value.song_name,
+      genres: value.meta,
+      singer: value.singer,
     };
-    const [affectedRows] = await Models.Song.update(updateFields, {
-      where: {
-        id,
-      },
+
+    const isAdmin = await Models.User.findOne({
+      where: { id: userId },
     });
-    if (affectedRows === 0) {
-      return res.status(http.NOT_FOUND.code).send({
+
+    if (isAdmin.dataValues.user_type === role.admin) {
+      const [affectedRows] = await Models.Song.update(updateFields, {
+        where: {
+          id,
+        },
+      });
+
+      if (affectedRows === 0) {
+        return res.status(http.NOT_FOUND.code).send({
+          success: false,
+          data: null,
+          message: http.NOT_FOUND.message,
+        });
+      }
+      const song = await Models.Song.findByPk(id);
+      await song.setGenres([]);
+      await song.addGenres(value.genres);
+
+      res.status(http.OK.code).send({
+        success: true,
+        message: messages.UPDATED,
+      });
+    } else {
+      res.status(http.FORBIDDEN.code).send({
         success: false,
         data: null,
-        message: http.NOT_FOUND.message,
+        message: http.FORBIDDEN.message,
       });
     }
-
-    res.status(http.OK.code).send({
-      success: true,
-      data: name,
-      message: http.CREATED.message,
-    });
   } catch (e) {
+    console.log(e);
     res.status(http.INTERNAL_SERVER_ERROR.code).send({
       success: false,
       data: null,
@@ -88,10 +131,8 @@ module.exports.UpdateSongData = async (req, res) => {
 module.exports.getSongData = async (req, res) => {
   try {
     const { page, pageSize } = req.query;
-
     const currentPage = parseInt(page, 10) || 0;
     const currentPageSize = parseInt(pageSize, 10) || 5;
-
     const offset = currentPage * currentPageSize;
     const limit = currentPageSize;
 
@@ -106,13 +147,15 @@ module.exports.getSongData = async (req, res) => {
         message: http.NOT_FOUND.message,
       });
     }
-    res.send({
+    res.status(http.OK.code).send({
+      success: true,
       data,
       message: http.OK.message,
     });
   } catch (e) {
     console.log(e);
-    res.send({
+    res.status(http.INTERNAL_SERVER_ERROR.code).send({
+      success: false,
       data: null,
       message: http.INTERNAL_SERVER_ERROR.message,
     });
@@ -122,17 +165,26 @@ module.exports.getSongData = async (req, res) => {
 module.exports.getSearchData = async (req, res) => {
   try {
     const { query } = req.query;
-    console.log("query", query, "typeof", typeof query);
+    const genre = await Models.Genre.findOne({
+      where: {
+        genre_name: query,
+      },
+    });
 
-    const data = await Models.Song.findAll({
-      where: Models.Song.sequelize.where(
-        Models.Song.sequelize.fn(
-          "JSON_CONTAINS",
-          Models.Song.sequelize.col("meta"),
-          query
-        ),
-        true
-      ),
+    const data = await Models.Genre.findAll({
+      where: {
+        id: genre.dataValues.id,
+      },
+      include: [
+        {
+          model: Models.Song,
+          through: {
+            attributes: [],
+          },
+          attributes: ["id", "song_name", "singer"],
+        },
+      ],
+      limit: 10,
     });
     if (!data) {
       return res.status(http.NOT_FOUND.code).send({
@@ -142,13 +194,15 @@ module.exports.getSearchData = async (req, res) => {
     }
     res.status(http.OK.code).send({
       success: true,
-      data,
+      data: data[0].dataValues.Songs,
       message: http.OK.message,
     });
   } catch (e) {
-    res.send({
+    console.log(e);
+    res.status(http.INTERNAL_SERVER_ERROR.code).send({
+      success: false,
       data: null,
-      message: "Something went wrong.",
+      message: http.INTERNAL_SERVER_ERROR.message,
     });
   }
 };
